# server/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CoachClientDetailResponse:
    properties:
      client:
        $ref: '#/definitions/dto.CoachClientSummaryResponse'
      dashboard:
        $ref: '#/definitions/dto.DashboardResponse'
    type: object
  dto.CoachClientSummaryResponse:
    properties:
      client_id:
        type: string
      email:
        type: string
      has_trained_this_week:
        type: boolean
      id:
        type: string
      last_workout_at:
        type: string
      nickname:
        type: string
      notes:
        type: string
      status:
        type: string
      total_workouts:
        type: integer
      weekly_volume:
        type: number
    type: object
  dto.CoachDashboardClientItem:
    properties:
      activity_bucket:
        type: string
      client_id:
        type: string
      email:
        type: string
      has_trained_this_week:
        type: boolean
      last_workout_at:
        type: string
      nickname:
        type: string
      weekly_volume_trend:
        items:
          $ref: '#/definitions/dto.VolumeDataPoint'
        type: array
    type: object
  dto.CoachDashboardResponse:
    properties:
      active_3_days:
        type: integer
      active_7_days:
        type: integer
      clients:
        items:
          $ref: '#/definitions/dto.CoachDashboardClientItem'
        type: array
      inactive_7_days:
        type: integer
      total_clients:
        type: integer
      trained_this_week:
        type: integer
    type: object
  dto.CoachInvitationResponse:
    properties:
      client_email:
        type: string
      coach_email:
        type: string
      coach_id:
        type: string
      coach_name:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  dto.DashboardResponse:
    properties:
      current_streak:
        type: integer
      last_week_volume:
        type: number
      muscle_distribution:
        items:
          $ref: '#/definitions/dto.MuscleVolumeData'
        type: array
      recent_prs:
        items:
          $ref: '#/definitions/dto.PRRecord'
        type: array
      volume_change_percent:
        type: number
      weekly_volume:
        type: number
      weekly_workouts:
        type: integer
    type: object
  dto.ExerciseResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      equipment:
        type: string
      id:
        type: string
      is_custom:
        type: boolean
      name:
        type: string
      name_en:
        type: string
      notes:
        type: string
      primary_muscles:
        items:
          type: string
        type: array
      secondary_muscles:
        items:
          type: string
        type: array
      tracking_type:
        type: string
    type: object
...
  dto.InviteClientRequest:
    properties:
      client_email:
        type: string
    required:
    - client_email
    type: object
...
  dto.MuscleVolumeData:
    properties:
      muscle:
        type: string
      sets:
        type: integer
      volume:
        type: number
    type: object
  dto.PRRecord:
    properties:
      achieved_at:
        type: string
      exercise_name:
        type: string
      id:
        type: string
      pr_type:
        type: string
      unit:
        type: string
      value:
        type: number
    type: object
  dto.PlanDayResponse:
    properties:
      exercises:
        items:
          $ref: '#/definitions/dto.PlanExerciseResponse'
        type: array
      id:
        type: string
      name:
        type: string
      sort_order:
        type: integer
    type: object
  dto.PlanDetailResponse:
    properties:
      created_at:
        type: string
      days:
        items:
          $ref: '#/definitions/dto.PlanDayResponse'
        type: array
      description:
        type: string
      goal:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.PlanExerciseResponse:
    properties:
      exercise:
        $ref: '#/definitions/dto.ExerciseResponse'
      id:
        type: string
      notes:
        type: string
      rest_seconds:
        type: integer
      sort_order:
        type: integer
      superset_group:
        type: integer
      target_reps:
        type: string
      target_sets:
        type: integer
      target_weight:
        type: number
    type: object
  dto.PushPlanRequest:
    properties:
      activate:
        type: boolean
      plan_id:
        type: string
    required:
    - plan_id
    type: object
...
  dto.VolumeDataPoint:
    properties:
      date:
        type: string
      volume:
        type: number
    type: object
  dto.WorkoutCommentRequest:
    properties:
      comment:
        maxLength: 2000
        minLength: 1
        type: string
    required:
    - comment
    type: object
  dto.WorkoutCommentResponse:
    properties:
      coach_id:
        type: string
      coach_name:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      workout_id:
        type: string
      workout_started_at:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Backend service for the strength training record app.
  title: Strength Tracker API
  version: "1.0"
paths:
...
  /coach/clients:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List coach clients
      tags:
      - coach
  /coach/clients/{clientId}:
    get:
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CoachClientDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get coach client detail
      tags:
      - coach
  /coach/clients/{clientId}/plans:
    post:
      consumes:
      - application/json
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: string
      - description: Push plan request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PushPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PlanDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Push plan to client
      tags:
      - coach
  /coach/clients/{clientId}/workouts:
    get:
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: string
      - description: Start datetime
        in: query
        name: date_from
        type: string
      - description: End datetime
        in: query
        name: date_to
        type: string
      - description: Exercise ID
        in: query
        name: exercise_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort_by
        type: string
      - description: Sort direction
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List coach client workouts
      tags:
      - coach
  /coach/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CoachDashboardResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get coach dashboard
      tags:
      - coach
  /coach/invite:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite client request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InviteClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CoachInvitationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Invite a client by email
      tags:
      - coach
  /coach/workouts/{workoutId}/comment:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workout ID
        in: path
        name: workoutId
        required: true
        type: string
      - description: Workout comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WorkoutCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkoutCommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add coach comment to workout
      tags:
      - coach
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"